pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hekate"
version = "1.0.0"
description = "Autonomous Multi-Agent Development System"
readme = "README.md"
license = {file = "LICENSE"}
maintainers = [
    {name = "Hekate Contributors"},
]
keywords = ["ai", "agents", "development", "automation", "orchestration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.10"
dependencies = [
    "redis>=5.0.1",
    "pyyaml>=6.0.1",
    "prometheus-client>=0.19.0",
    "pytest>=7.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]
docs = [
    "sphinx",
    "myst-parser",
]

[project.scripts]
hekate = "hekate.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hekate"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q"
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]